requirements/base.txt
django==4.1.7
djangorestframework==3.14.0
django-environ==0.9.0

# Websocket
channels==4.0.0
daphne==4.0.0
channels-redis==4.0.0
